Feature/queue cancel queue ticket

+++ src/main/java/com/swp/autocarwash/booking/dto/response/CreateBookingResponse.java
@@ -26,5 +26,5 @@ public class CreateBookingResponse {
     private Long bookingId;
     private String status;
     private BigDecimal totalAmount;
-    private List<Integer> slotIds;
+    private List<Long> slotIds;
 }
